requirements.txt
google-generativeai
datasets
pandas
gradio
python-dotenv
huggingface_hub
